Changed ulong params for RenderSystem::setStencilBufferParams to uint32.

+++ OgreMain/include/OgreRenderSystem.h
@@ -788,7 +788,7 @@ namespace Ogre
             and the inverse of them will happen for back faces (keep remains the same).
         */
         virtual void setStencilBufferParams(CompareFunction func = CMPF_ALWAYS_PASS, 
-            ulong refValue = 0, ulong mask = 0xFFFFFFFF, 
+            uint32 refValue = 0, uint32 mask = 0xFFFFFFFF, 
             StencilOperation stencilFailOp = SOP_KEEP, 
             StencilOperation depthFailOp = SOP_KEEP,
             StencilOperation passOp = SOP_KEEP,
